Auth, Sync and Get api's are all working correctly.
On the other hand Insert and Update are not. I'm still not decided on the format so it may change.

NOTE: to get the Sync API to work, a Syncable abstract class was introduced which all classes extend from, adding an 'added' and 'last_modified' field to those classes. Thus this revision requires a database rest when moving from the previous version.