Completed the GoogleAdapter and the APi BookProfile save.
This commit introduces the requirement for django-tagging, which needs source code modification to get working.

Find the installation path, go to tagging > models.py and on line478 change the PositiveIntegergerField() to be a CharField() (give it a max_length of 64), as this makes it work for UUIDs as well as regular primary keys.