Completely revamped the user interface, basically made the collection and book lists look a whole lot nicer.
Set up layout for moving the unsorted books list into the library_detail view.

Modified the add_books view to use the same list style and use icons to represent whether a book was added to the collection or not.

Add a "quick add" feature to bookshelves so that books maybe added specifically to that bookshelf.